fix: JPEGの場合にGPS情報を削除する
