+++ app/uploaders/figure_uploader.rb
@@ -1,9 +1,8 @@
 class FigureUploader < CarrierWave::Uploader::Base
   include CarrierWave::Vips
+  include StripGps
 
   storage :file
-
-  process :strip_gps, if: :jpeg?
 
   def store_dir
     "uploads/#{model.class.to_s.underscore}/#{mounted_as}/#{model.id}"
@@ -56,19 +55,6 @@ class FigureUploader < CarrierWave::Uploader::Base
     overlay_img_rgb = overlay_img.extract_band(0, n: 3)
     overlay_img_with_alpha = overlay_img_rgb.bandjoin(alpha)
     src_img.composite(overlay_img_with_alpha, :over, x:, y:)
-  end
-
-  def strip_gps
-    cache_stored_file! if !cached?
-    exif = MiniExiftool.new(current_path)
-    exif.GPSLatitude = nil
-    exif.GPSLongitude = nil
-    exif.GPSAltitude = nil
-    exif.save
-  end
-
-  def jpeg?(carrier_wave_sanitized_file)
-    MimeMagic.by_path(carrier_wave_sanitized_file.path).type == "image/jpeg"
   end
 
   def gif?(carrier_wave_sanitized_file)
